Contact us page
in css .map if you add
position: absolute;
top:50px;
right: 300px;
width: 500px;
we are able to align the map but map will not remain responsive at any
resolution. so I personally suggest you to make that as it is despite of
align that into a particulr position.  or if any of you have any idea to
to align that but map should remain resposive could make the changes.
cheers